requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
passlib[bcrypt]
argon2-cffi
pydantic
httpx
pytest
email-validator
redis
aio_pika
alembic
